fix: remove invalid worker type imports
