+++ packages/engine/src/simulation/pathfinding/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './advancedPathfinding';
+export { aStar } from './algorithms';
