+++ packages/engine/src/simulation/workerSimulation.ts
@@ -1,14 +1,14 @@
 import type { SimResources } from '../index';
 import type { SimulatedBuilding } from './buildingSimulation';
 import type { Citizen } from './citizenBehavior';
-import {
-  workerSystem,
-  type WorkerProfile as SystemWorkerProfile,
-  type JobAssignment,
-} from './workerSystem';
-import type { WorkerProfile as SystemWorkerProfile, JobAssignment } from './workers/types';
+import { workerSystem } from './workerSystem';
+import type {
+  JobAssignment,
+  WorkerProfile as SystemWorkerProfile,
+  JobRole,
+  WorkerProfile,
+} from './workers/types';
 import type { GameTime } from '../types/gameTime';
-import type { JobRole, WorkerProfile } from './workers/types';
 import {
   calculateWageAdjustment,
   checkCareerProgression,
